Rename executor/internal/utils to util
