+++ executor/internal/util/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "utils",
+    name = "util",
     srcs = ["partition_table.go"],
-    importpath = "github.com/pingcap/tidb/executor/internal/utils",
+    importpath = "github.com/pingcap/tidb/executor/internal/util",
     visibility = ["//executor:__subpackages__"],
     deps = [
         "@com_github_pingcap_errors//:errors",
